+++ app/src/components/grid/stdGrid.tsx
@@ -29,7 +29,7 @@ import GridLoadingOverlay from "./gridLoadingOverlay";
 import "./style.css";
 
 // duckGrid Folder
-import duckGridDataSource from "../duckGrid/duckGridDS";
+import duckGridDataSource from "../duckGridDS/duckGridDS";
 import CustomCountBar, {
   CustomFilterModelBar,
   CustomWaterMarkBar,
